requirements.txt
PyYAML==3.12
requests==2.12.1
six==1.10.0
slackclient==1.0.2
websocket-client==0.37.0
